moved functions word case
